Revised URL creation login in RestClient
The URL creation in RestClient was creating errors for the “verify user” call because that specific request requires a ? in its endpoint.

The existing logic added a ? to every endpoint, followed by the query string (if one exists).

This created an issue for “verify user” where two question marks were in the endpoint, resulting in calls not working correctly.

This has been revised to only add a quesiton mark if there are $params. If there are no $params, the URL will simply be the entity/endpoint.

If $params exist, if the entity/endpoint includes a ?, then the $params will be appended using an ampersand. If the entity/endpoint does not include a ?, then the $params will be appended using a ?.

The result is a smarter URL creation logic that works for entity/endpoint calls that do include a question mark - as far as I’m aware, there is only the “verify user”, and only applicable to pre-live calls. But still creates an issue for any developer doing testing with that call.